+++ packages/client/src/views/Duplicates/DuplicateWarning.tsx
@@ -11,7 +11,7 @@
  */
 import React from 'react'
 import styled from 'styled-components'
-import { Warning } from '@opencrvs/components/lib/interface'
+import { Alert } from '@opencrvs/components/lib/Alert'
 import { Query } from '@client/components/Query'
 import { errorMessages } from '@client/i18n/messages/errors'
 import { useIntl } from 'react-intl'
@@ -54,7 +54,7 @@ export function DuplicateWarning({
               const duplicateQuery = data?.[`duplicate${idx}`]
               if (duplicateQuery?.registration?.trackingId) {
                 return (
-                  <Warning
+                  <Alert
                     label={intl.formatMessage(errorMessages.duplicateWarning, {
                       trackingId: duplicateQuery.registration.trackingId
                     })}
