+++ packages/client/src/views/RegisterForm/RegisterForm.tsx
@@ -27,9 +27,9 @@ import { BackArrow } from '@opencrvs/components/lib/icons'
 import {
   EventTopBar,
   IEventTopBarProps,
-  IEventTopBarMenuAction,
-  Warning
+  IEventTopBarMenuAction
 } from '@opencrvs/components/lib/interface'
+import { Alert } from '@opencrvs/components/lib/Alert'
 import { ResponsiveModal } from '@opencrvs/components/lib/ResponsiveModal'
 import { Spinner } from '@opencrvs/components/lib/Spinner'
 import { Container, BodyContent } from '@opencrvs/components/lib/Content'
@@ -790,7 +790,7 @@ class RegisterFormView extends React.Component<FullProps, State> {
                         >
                           {isFileSizeExceeded(declaration) &&
                             isDocumentUploadPage && (
-                              <Warning
+                              <Alert
                                 label={intl.formatMessage(
                                   constantsMessages.totalFileSizeExceed,
                                   {
