+++ packages/components/src/interface/index.ts
@@ -31,4 +31,3 @@ export * from './LocationSearch/LocationSearch'
 export * from './ErrorToastNotification'
 export * from './TableView'
 export * from './Pill'
-export * from './Warning'
